Create types file and udpate docs to tsdoc
